+++ lib/services/doorman.dart
@@ -1,33 +1,32 @@
 import 'dart:async';
 
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:fuks_app/generated/google/protobuf/empty.pb.dart';
 import 'package:fuks_app/generated/doorman.pbgrpc.dart';
-import 'package:fuks_app/services/doorman_cert.dart';
 import 'package:grpc/grpc.dart';
+import 'package:grpc/grpc_or_grpcweb.dart';
 
-// final _channel = GrpcOrGrpcWebClientChannel.toSingleEndpoint(
-//   // host: '192.168.0.125',
-//   host: 'doorman.local',
-//   port: 50051,
-//   transportSecure: true,
-// );
-
-final _channel = ClientChannel(
-  'doorman.local',
+final _channel = GrpcOrGrpcWebClientChannel.toSingleEndpoint(
+  host: '192.168.0.152',
+  // host: 'doorman.local',
   port: 50051,
-  options: ChannelOptions(
-    credentials: ChannelCredentials.secure(
-      certificates: doormanCert.codeUnits,
-      onBadCertificate: (certificate, host) {
-        debugPrint('certificate: ${certificate.issuer}');
-        debugPrint('host: $host');
-        return host == 'doorman.local:50051';
-      },
-    ),
-  ),
+  transportSecure: false,
 );
+
+// final _channel = ClientChannel(
+//   'doorman.local',
+//   port: 50051,
+//   options: ChannelOptions(
+//     credentials: ChannelCredentials.secure(
+//       certificates: doormanCert.codeUnits,
+//       onBadCertificate: (certificate, host) {
+//         debugPrint('certificate: ${certificate.issuer}');
+//         debugPrint('host: $host');
+//         return host == 'doorman.local:50051';
+//       },
+//     ),
+//   ),
+// );
 
 final doorman = DoormanServiceWithToken();
