+++ lib/ui/pages/home/access.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:fuks_app/generated/doorman.pb.dart';
 import 'package:fuks_app/services/doorman.dart';
-import 'package:fuks_app/ui/dialogs/error.dart';
+import 'package:fuks_app/ui/widgets/error.dart';
 import 'package:fuks_app/ui/pages/home/access_timer.dart';
 import 'package:fuks_app/ui/widgets/constrained_list_view.dart';
 import 'package:undraw/undraw.dart';
