+++ lib/ui/widgets/error.dart
